Validators refactors
